+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- Endpoint for customer creation       (POST shopgate/v2/customers)
+### Added
+- Endpoint for customer creation (POST shopgate/v2/customers)
 
 ## [3.3.0] - 2018-12-05
 ### Added